Fix random melee weapon.
Note: most melee weapons are not working correctly, in that the melee kill counts as getting a kill with the
primary weapon and consequently promotes the player. Also, it seems that getting a kill with melee weapon as
last level primary weapon no longer promotes the player, ending the game.

Will need to fix both issues.